+++ src/Ormolu/Fixity/Internal.hs
@@ -214,6 +214,11 @@ defaultModuleReexports =
             "Optics.IxTraversal",
             "Optics.Traversal"
           ]
+      ),
+      ( "Test.Hspec",
+        NE.fromList
+          [ "Test.Hspec.Expectations"
+          ]
       )
     ]